Change type of level and skill point

+++ Scripts/MMOGame/Database/MySQL/MySQLDatabase_Guild.cs
@@ -33,9 +33,9 @@ namespace MultiplayerARPG.MMO
             if (reader.Read())
             {
                 result = new GuildData(id, reader.GetString("guildName"), reader.GetString("leaderId"), defaultGuildRoles);
-                result.level = reader.GetInt32("level");
+                result.level = (short)reader.GetInt32("level");
                 result.exp = reader.GetInt32("exp");
-                result.skillPoint = reader.GetInt32("skillPoint");
+                result.skillPoint = (short)reader.GetInt32("skillPoint");
                 result.guildMessage = reader.GetString("guildMessage");
 
                 reader = ExecuteReader("SELECT * FROM guildrole WHERE guildId=@id",
